refactor(pipettes): reduce some of the warnings generated in the pipettes project
There are quite a lot of warnings, related to different things. I have tried to reduce warnings that
were unrelated to the HAL layer of things. Some of these files were touched in the FreeRTOS PR, but
I plan to open a separate PR for that once the motor controller task stuff is in a good place. There
were a few CAN related warnings that I did not know how to resolve.

RET-191