cmake presets: update to schema 2, add builds (#124)
Cmake preset schema version 2, introduced in CMake 3.20, allows for
"build presets" that are applied when invoking cmake --build. Like
configure presets, they let you specify options that you'd otherwise
have to remember to give cmake every time, including most significantly
directories. With the introduction of a second cross configuration
because of a different microcontroller architecture, it's getting to be
a bit of a pain to remember.

In addition, we'll specify a magic environment variable that the cmake
configure code can use as a sign that a preset was used.
This is important because otherwise you can still just call cmake
without specifying a preset, and that will generate trash files all over
the place.

Finally, we'll also specify in our presets the kind of architecture
we're using and only add arch-specific subdirectories if they're
relevant, which should help prevent a problem where you can just call
cmake --build with whatever target you want.

All in, this should require devs to specify presets at configure time
and prevent them from incorrectly using cmake, at least in this specific way.