refactor(*): parametrize motor hardware iface
Refactor the way the motor control system (what emits step and direction
pulses) interacts with the hardware layer.

The current implementation had no single source of truth for what pins
did what and how to interact with them, leading to a lot of duplicated
code and a tough time following the way motor commands went from coming
in on the canbus to going out on gpio lines.

Instead of using static c functions to implement setting pin state, we
can make them one implementation of a c++ pure virtual base class, which
we can give in place of the pin configurations to the motion controller
and also provide to the motor interrupt handler. This makes the full
loop from getting a motor command to calling some function that sets a
pin live in the core.

This commit creates and implements this interface for firmware and
reshuffles a bunch of code to integrate it into the three end
executables and rehome some of the implementations in a new common
subdirectory. It also removes some templatization from things like the
motor interrupt handler, which were only templatized over their queue
implementation, which is actually invariant since we're using POSIX
freertos in test and simulator.