fix(pipettes): make timer interrupt 100KHz
The timer interrupt for motor control in the pipettes subproject was going too fast. The result is
that the CAN task could never finish its initialization process, and thus, the timer interrupt for
FDCAN would never get called when a message was sent from python. For now, we are setting the
interrupt in pipettes to be 100KHz. We may want to change the interrupt to match the minimum
requirement for all the stepper motors in the future.