feat: add python bindings
We face this problem of coordinating repos, where the python side of
things needs to know many constants and message data structures that
need to match c++. It's really annoying to keep these synchronized, and
really annoying to do cross-repo work. The solution is to move the
bindings to the same place as the firmware, and (eventually) make a
single source of truth for the messages by having the python generate
the c++ (since that code is written already).

This adds a new directory python/ containing a new python package
opentrons_ot3_firmware, which contains the section of opentrons_hardware
that just contains static object definitions for messages and also some
constants, along with its tests, and enough cmake python stuff to run
the tests and format.