fix(gantry): x direction inversion (#219)
X homes to the physical right of the robot, and since we want to keep
its positions positive and home to 0, that means that positive motion is
right to left. That means we need to invert the sense of the direction
pin, which is now different from Y, so we need separate settings.