feat(can): Node IDs for left and right pipette
When there are two pipettes attached, they need to have different node
IDs so they can both talk on the bus. The thing that makes them
different from each other is which mount they're on. We therefore can
use a different node id depending on the mount, which can be sensed from
a voltage expressed by a resistor net split between pipette and head
boards.

This commit only creates the new IDs and switches to using them. The pipette
application firmware and bootloader listen to left, arbitrarily.

The next step is to bring up the ADC to sense the voltage on the
resistor net and store its value so we know who we are.