feat(pipettes): Parametrize core tasks by node id
The node id needs to be specified at boot. This PR won't contain the
code that determines the value of the node id from the ADC, but if we're
going to have a bunch of defaults we might as well have just one
default.

Configure the can task reader and writer to get their values from when
main starts the stack.

In the simulator, specify the MOUNT env var to "left" or "right" to use
the appropriate one (default - left).

In the firmware, one TODO is left which sets the node id to left.