feat(can): Node IDs for left and right pipette (#230)
When there are two pipettes attached, they need to have different node
IDs so they can both talk on the bus. The thing that makes them
different from each other is which mount they're on. We therefore can
use a different node id depending on the mount, which can be sensed from
a voltage expressed by a resistor net split between pipette and head
boards.

The node id needs to be specified at boot. This PR won't contain the
code that determines the value of the node id from the ADC, but if we're
going to have a bunch of defaults we might as well have just one
default.

Configure the can task reader and writer to get their values from when
main starts the stack.

In the simulator, specify the MOUNT env var to "left" or "right" to use
the appropriate one (default - left).

In the firmware, one TODO is left which sets the node id to left.