refactor(head,common): commonize tool detect
The tool detection will also be needed on the pipette. In addition to
therefore needing the tool detection code to live in common/ somewhere,
the pipettes also need to interact with the code in a different way -
they don't care what kind of tool they are (or maybe do only as a
debug), just what carrier they're on. In the same way, the head doesn't
really care what carrier it reads, since it knows which pins it's
looking at, just which tool it is. So we can provide separate filtering
options for each and let each use of the code become simpler.

Unfortunately ranges and views just don't work in clang at all because
of https://github.com/llvm/llvm-project/issues/44178 so even though it
all compiles correctly we have to do a pretty ugly hack that removes
functionality when linting. It would be nice to not have this.