refactor(head,common): commonize adc support
There's some ADC stuff that's different between devices (well, a lot of
ADC stuff) but the reading->mv transition isn't, so we might as well
commonize it.

The core element here is an ADC channel, which is templatized with the
conversion variables. That lets it do some nice conversions, and with
some required implementations in the child we can have a nice
generalizable interface.

There isn't an ADC interface in common, because that's going to go a
bunch of different places - the head has one, for instance, that
specifically has gripper, z, and a channels - nothing else needs that.
It relies on the common channel implementation, and the simulation and
test versions override as appropriate.