feat(pipette): Detect mount at boot and assign node id (#246)
- refactor(head,common): commonize tool detect

The tool detection will also be needed on the pipette. In addition to
therefore needing the tool detection code to live in common/ somewhere,
the pipettes also need to interact with the code in a different way -
they don't care what kind of tool they are (or maybe do only as a
debug), just what carrier they're on. In the same way, the head doesn't
really care what carrier it reads, since it knows which pins it's
looking at, just which tool it is. So we can provide separate filtering
options for each and let each use of the code become simpler.

Unfortunately ranges and views just don't work in clang at all because
of https://github.com/llvm/llvm-project/issues/44178 so even though it
all compiles correctly we have to do a pretty ugly hack that removes
functionality when linting. It would be nice to not have this.

- refactor(head,common): commonize adc support

There's some ADC stuff that's different between devices (well, a lot of
ADC stuff) but the reading->mv transition isn't, so we might as well
commonize it.

The core element here is an ADC channel, which is templatized with the
conversion variables. That lets it do some nice conversions, and with
some required implementations in the child we can have a nice
generalizable interface.

There isn't an ADC interface in common, because that's going to go a
bunch of different places - the head has one, for instance, that
specifically has gripper, z, and a channels - nothing else needs that.
It relies on the common channel implementation, and the simulation and
test versions override as appropriate.

- feat(pipettes): Detect current mount at boot

Based on the voltage present on the tool sense pin and using the ADC and
mount detection utilities previously added to common/, we can now pretty
easily decide which mount we're on and assign ourselves a canbus id
based on that.

- feat(bootloader): add pipette mount detection

This unfortunately reimplements the mount detection algorithm in C but
it will work just fine (and is checked by unit tests). Bootloaders
should now properly decide which mount they're on.

- fix(bootloader): allow flashing pipettes

This flash command had the wrong openocd config value