refactor: generate canbus definitions manually (#255)
Commit the outputs of the header generation, move the python source to
the monorepo,  and add a script and cmake target that will generate
headers from a separately-checked-out monorepo.

The workflow for dealing with canbus definitions mirroring between
python and c/++ was kind of a pain - not here, but in the monorepo. The
monorepo's use of pipenv really isn't friendly to this sort of thing,
and updating the version of this repo used in the monorepo's python was
just as manual as anything this commit does.

After this change, check out the monorepo to a sibling directory and run
cmake --build --preset=<anything> --target update-headers and you'll
write the changes locally.