feat(head): tolerate unknown tool ids
The tool identification voltage ranges will sometimes not quite be
right, and the values coming out of the ADC can bounce around slightly.
This can combine to have the head spam the canbus with tool detection
notification change messages bouncing between some normal value and
undefined_tool. In practice, if these values bounce around like this,
it's not reflecting anything physical - so let's filter it out by
latching away undefined_tool, and not allowing the state to go from
valid->invalid since undefined_tool is just an artifact of the way we
categorize things.