feat: add version int and git sha to get device info (#280)
- Add version generation
The version will be a monotonically increasing number (it's much easier)
and the responses will build in the git shortsha. This can then be sent
over in system info.

- Send version and sha in device info

The version will just be a monotonically increasing integer like chrome
because these things are not things that users ever need to install. To
handle the common case where these are all dev builds, we also package
the sha of the current commit.

- add some flags for typical problems

It's really useful to know whether a build was a dev build or from CI,
and whether it actually was built only from the content at the specified
commit hash, and (without looking it up) whether the commit was actually
the version specified

- use version in bootloader