fix(build): proper deps for combined hex files
Now that we have bootloaders, the thing we flash when we run a
`project-flash` target is a hex file created by a python script
combining an application hex file and a bootloader hex file.

This was done in cmake by having custom commands+targets for the
application and hex files, and then a custom command+target for the
combined hex called `target-hex-image`.

The custom command for `target-hex-image` had a DEPENDS for the custom
target portion of the application and bootloader hexes. Unfortunately,
per https://cmake.org/cmake/help/v3.21/command/add_custom_command.html,
a DEPENDS argument that is a custom target creates a target level
dependency "to make sure the target is built before any target using
this custom command"; but because the DEPENDS arg is a custom target and
_not_ a library or executable, it does _not_ create a file-level
dependency to cause the command to rerun when its deps change.

That means that `$TARGET-image-hex` and `$TARGET-image.hex` (which is
the one used by the `$TARGET-flash` command) was stealthily not being
rebuilt when you changed source code, even though the actual application
was. This meant changes would not apply until you used the
`$TARGET-debug` target, which does not use the combined hex image.

The way to fix this (and what this commit does) is to not only have a
dependency on the custom targets for the individual application/binary
hexes, but also have direct file dependencies on their outputs; this
creates the proper deps and now the combined image gets rebuilt properly.