Ret512 encoder firmware c configuration (#291)
* motor encoder configuration files

Following functions added
-motor encoder start timer
-motor encoder stop timer

Created a header file with the above functions

* add timer2 and timer3 configuration

In order to activate the motor encoders, we need to configure the timers for Z axis and A axis. These timers are tim2 and tim3. These timers are specifically for just motor encoders Z axis and A axis.

* fixed gpio pin declaration

* enable the HAL_RCC_TIMERS

Enable the RCC TIMER CLK for TIMER2 and TIMER3.

TIMER2 for the Motor Encoder A
TIMER3 for the Motor Encoder B

* commend added

add commends

* removed unnecessary lines

clean up on the encoder ZA encoder configuration.

* removed stbool and comments

stbool is not being used, so removed it from the header file.

* changed htim2 period to the max

the motor encoder pins are inputs and don't need to be reset. We can just reset the count by resetting the counter value to zero.

* deleted these files

these files are not currently being used and don't need them at the moment.

* i removed cmake command by mistake

should fix the compiler issue

* added encoder gpio pins

Just to be safe I'm adding these gpio pins. I'm not sure why this worked before let's make sure to initialize the pins for each motor encoder on the head.

* fix conflict changes