feat(pipettes): allow different pipette types
You should now be able to select and build code with the correct pipette type. Since the main change
here is mostly around `pipettes-core` I decided to keep the executable as `pipettes` and choose
which core to load based on an environment variable. Unsure if this is the best approach but I was
trying to minimize the copy/pasting required.

Closes #RET-663