refactor(motor-control): add LMS to move status
For the move status reporter to report positions in physical units, it
needs to know the translation between the positions it gets given (in
steps) from the motor interrupt handler, and those physical units.
That's in the linear motion system object, which the reporter handler
needs to have. Thread that through the type system.