feat(*): real position in move-complete (#304)
* refactor(motor-control): add LMS to move status

For the move status reporter to report positions in physical units, it
needs to know the translation between the positions it gets given (in
steps) from the motor interrupt handler, and those physical units.
That's in the linear motion system object, which the reporter handler
needs to have. Thread that through the type system.

* refactor(motor-control): add um_per_step to lms

To avoid frequently dividing, we can have a single calculation that gets
the um_per_step, the inverse of steps_per_mm (but in micrometers, since
unlike steps_per_mm this will almost always if not literally always be <1mm).

* feat(*): send current position in um after move

We've always been sending the current position back after a move
completes, but it's also always been in steps. That isn't a useful unit
when the host doesn't know the steps/mm of each device.

Instead, let's use our knowledge of the mechanical configuration to
translate the current position from steps to micrometers before sending
it back over the canbus.

* fix(head): fix simulator motor bindings

Revealed by an integration test, no less - the simulator wasn't binding
the motor interfaces correctly.

* fix(motor-control): set pos to 0 earlier in home

Otherwise, the position won't get reported correctly.