refactor(common): dynamic and reusable timers
The timer had its period as a template argument, which was nice but
means it can't ever be changed once created. Change that to a member
variable.

Also, add C++ member wrappers for the most of the rest of the freertos
timer api so the timer can be reset, change its period, etc etc.