feat(pipettes): Add I2C poller
The polling functionality of the I2C handling task was implemented in a
blocking fashion (in its own task and freertos compliant, but the I2C
task could not do other work while polling). This is a problem if we
want to add continuous polling, which we need for triggering GPIO pins
based on changes in the sensed values.

Instead, let's make a new task (that doesn't really do much when it
runs) that serves as a home for a whole ton of timers, which fire at the
period requested for polling and use the I2C writers to send and receive
values on the bus. This can support limited-count polling (through a
fairly ugly curried-recursion method) as well as continuous polling.