im so sorry
This commit contains two major pieces of work.

First, I was finally unable to resist the siren song of getting the i2c
stuff into its own subproject. I can only blame the crew, who did not
lash me to the mast tightly enough.

That comes with new github actions workflows, dependencies in the
pipette on the new subdirectory, and a whole bunch of renames in a whole
bunch of places.

Second, we finally found the problem that we were debugging: if you
memcpy std::functions, that obviously doesn't trigger their copy ctor,
so the memcpyd-from function gets destroyed when it leaves the scope
where it was memcpyd, and there goes its closure object. This is going
to be a consistent problem with anything with a non-trivial dtor that
goes through a message buffer.

The solution is to instead pass around queue handles that we can write
responses to. This is in itself kind of a pain since we can't use std
functions! We kind of have to bodge in manual c-style closures with
reinterpret cast, but we can at least protect the building of those
closures with some nice template concepts.

At this point, there also really didn't seem like a point keeping the
various different kinds of i2c messages so now there's just transact and
transactionresponse.

The i2c subdirectory compiles and passes the tests for i2c_task and
i2c_writer; it probably doesn't compile or pass the test for poller, but
a) it's kind of cool that the includes are clean enough that you can
just not run the poller tests and it's fine, and b) yeah no kidding. It
also For Sure will not be able to compile the pipettes until the sensor
and eeprom classes are adapted.

add memcpy test

add nicer comments

poll tests pass

more poller tests pass

all poller tests pass!