fix(pipettes): correctly assign i2c buses
The hardware for the pipettes actually assigns all the sensors to the
i2c1 bus and the eeprom on the i2c3 bus, no matter the kind of pipette.
That means that we don't need the single/multiple mains and tasks, since
we're always using both buses, and we need to swap some other stuff
around.

After this commit, messages and data can be observed on i2c when
querying sensors.