feat(pipette): add sensor bindings (#316)
* feat(pipettes): add sensor binding messages

Calibration and liquid sense require the pipette to pulse the sync line
active when sensor conditions exist. This commit adds a message to tell
the pipette to configure the sync line to be controlled by one of its sensors.

* feat(pipettes): add sync out pin handling

* refactor(common): dynamic and reusable timers

The timer had its period as a template argument, which was nice but
means it can't ever be changed once created. Change that to a member
variable.

Also, add C++ member wrappers for the most of the rest of the freertos
timer api so the timer can be reset, change its period, etc etc.

* refactor(pipettes): add i2c-transact method

This lets you read and immediately write a value to an address, which is
pretty useful for reading specific registers.

* feat(pipettes): Add I2C poller

The polling functionality of the I2C handling task was implemented in a
blocking fashion (in its own task and freertos compliant, but the I2C
task could not do other work while polling). This is a problem if we
want to add continuous polling, which we need for triggering GPIO pins
based on changes in the sensed values.

Instead, let's make a new task (that doesn't really do much when it
runs) that serves as a home for a whole ton of timers, which fire at the
period requested for polling and use the I2C writers to send and receive
values on the bus. This can support limited-count polling (through a
fairly ugly curried-recursion method) as well as continuous polling.

* refactor(pipettes): poll functionality moves

Now that the new poller tasks exist, we can remove the i2c writer's
capability to handle polling and move all the tasks dependent on it over
to the new poller.

* refactor(pipettes): easier message specs for i2c

Now you can prep a buffer or do the big integer thing, whichever is
easier at the time.

* tests compile with changes

* restore ability to send ints via i2c

Sometimes it's just way easier.

* fix rest of i2c task tests

* sensors use ints again

* add timer interface and mock timer

* template timer in poller task

add poll task tests for limited polls

add tests for continuous polling

add poll frontend tests and fix the bugs it finds

* add timer to pipette sim freertos config

and fix some cmakelist usage

* refactor: move gpio cpp defs to common

We had a fairly rich set of functions and datatypes for defining and
interacting with gpios (that are configured elsewhere) but they were all
in motor-control. We now need them in sensors. Move this all to common.

* pipettes: bind sync line hardware into sensors

This will let the sensors change the sync line based on sensed
conditions. The hardware is now plumbed into the capacitive sensor interface.

* pipettes: bind sensor outputs from can

When the sensor tasks get the appropriate message, they will set up a
new (or modify an ongoing) poll that will take the appropriate output
action - for now, report or ping the sync line.

* fix some lint issues

* adapt pressure sensor to poller

pressure sensor uses transact()

* Delete gpio_interface.hpp

* fixes from testing

* more timer fixes

* im so sorry

This commit contains two major pieces of work.

First, I was finally unable to resist the siren song of getting the i2c
stuff into its own subproject. I can only blame the crew, who did not
lash me to the mast tightly enough.

That comes with new github actions workflows, dependencies in the
pipette on the new subdirectory, and a whole bunch of renames in a whole
bunch of places.

Second, we finally found the problem that we were debugging: if you
memcpy std::functions, that obviously doesn't trigger their copy ctor,
so the memcpyd-from function gets destroyed when it leaves the scope
where it was memcpyd, and there goes its closure object. This is going
to be a consistent problem with anything with a non-trivial dtor that
goes through a message buffer.

The solution is to instead pass around queue handles that we can write
responses to. This is in itself kind of a pain since we can't use std
functions! We kind of have to bodge in manual c-style closures with
reinterpret cast, but we can at least protect the building of those
closures with some nice template concepts.

At this point, there also really didn't seem like a point keeping the
various different kinds of i2c messages so now there's just transact and
transactionresponse.

The i2c subdirectory compiles and passes the tests for i2c_task and
i2c_writer; it probably doesn't compile or pass the test for poller, but
a) it's kind of cool that the includes are clean enough that you can
just not run the poller tests and it's fine, and b) yeah no kidding. It
also For Sure will not be able to compile the pipettes until the sensor
and eeprom classes are adapted.

add memcpy test


* refactor: move sensor tests to own home

The includes were in their own home but the test source wasn't. Now they
are, hooray!

Code is on the way to being adapted to message responses, and the
capacitive sensor tests build (don't pass though).

finish cap sense tests

enviro sensor compiles



Co-authored-by: Caila Marashaj <caila.marashaj@opentrons.com>