refactor(sensors): capacitive sensor fixes
The capacitor sensor needed its behavior changed a little to properly
re-offset its input with capdac and handle the input parsing a little
better. The conversion values are left-aligned and are signed (relative
to the current offset), and the offset needs to be changed when values
start drifting away from 0. This may need future work to lo-pass the
offset changes - they can be really hit by noise - but for now, it works.