feat(sensors): add threshold autodetect
The capacitive sensor on the pipettes is used to detect collisions
between a conductive probe (or tip) and the deck by triggering when the
sensed capacitance changes. To make this more accurate and resilient, we
want to get the sensor very close to - but not touching - the deck, save
the capacitance at that time as a baseline, and then trigger when the
capacitance diverges from that baseline.

This could be accomplished by having the host request a sensor reading,
calculate a threshold, and then send the threshold over; but it could
also be accomplished by having the host say "go take a reading and set
your threshold to that + 1", with a couple fewer canbus messages, which
is what this does.