refactor(*): commonize task starters (#327)
The task starter objects that we have, one for each task, serve two
purposes:
1. Hold the types that tests can't use
2. Hold the data storage for tasks on the free store

These are both useful but they're also both generic. By making a single
generic task starter, we can get rid of the other ones.

That will be a lot easier if all the tasks are a little better behaved,
such as
- Only have a single class template argument (the queue) that we can
then fill in with a freertos queue
- Have all the template arguments to their run methods deducable from
passed arguments

Once that's true, we can entirely get rid of the dedicated task builders.