refactor(sensors): fix capacitive sensor handling (#324)
* fix(pipettes): correctly assign i2c buses

The hardware for the pipettes actually assigns all the sensors to the
i2c1 bus and the eeprom on the i2c3 bus, no matter the kind of pipette.
That means that we don't need the single/multiple mains and tasks, since
we're always using both buses, and we need to swap some other stuff
around.

After this commit, messages and data can be observed on i2c when
querying sensors.

* drop to 100khz bus speed

because 400khz is frankly unreasonable over a ribbon wire harness

* organize fdc config in arrays

somewhat better than prepacked ints - much easier to pack

* refactor(sensors): capacitive sensor fixes

The capacitor sensor needed its behavior changed a little to properly
re-offset its input with capdac and handle the input parsing a little
better. The conversion values are left-aligned and are signed (relative
to the current offset), and the offset needs to be changed when values
start drifting away from 0. This may need future work to lo-pass the
offset changes - they can be really hit by noise - but for now, it works.

* fix(*): drive sync out high

It can't be left floating on the output side because it's buffered

* fix(sensors): actually set bind_sync

* wrong setting for sync in head

* push/pull sync out

* respect all message stop conditions

* clean up sync

* properly propagate lint targets again

The parent scope in the individual target functions stopped working
because they were functions, which get scopes. Add another layer of
upward propagation to get the lint target working in pipettes.

* fix lint problems now that lint works again