chore(can): Add can bit timing calculations
Can bit timings are a bit in depth but aren't actually that complex,
especially as parametrized on the stm32 canbus modules. They depend on
the clock you're giving them, what you want the time quantum to be, what
you want the bitrate to be, and when you want to sample. Given all those
you can calculate numbers, and with c++ consteval it's easy to do it at
compile time with static asserts for out of range values (including
narrowing!). It's a bit messy because it all needs to be templates,
though.

The values specified here are also different from those previously
hardcoded. The values previously hardcoded
- Don't match the time quantum or sample point used by the host (they
have a much longer quantum, for instance) and
- Don't match the sample point used by the host, and
- Have different TQs across the pipettes and G4s for some reason, and
- Consequently have a different bitrate (slightly) on the pipettes, and
- Even though we're not doing bitrate switching, have different values
for data segment bit timings.

This should fix all that. The values will still be a little different on
each core just by virtue of the canbus getting fed different clocks, but
they should be a lot closer now - the smaller you can make the time
quantum, the more accurate you can be by wall clock.