chore(can): update timings to match linux
Linux uses significantly different canbus bit timing parameters from the
microcontrollers. This may not be a problem but it also certainly may be
- the sync jumps are very different and may lead to very rapid
synchronization loss; and the pipettes are wildly different from the rest of
the micros.

This fixes the above by setting the values to be the same as linux.