chore(can): generate can timings (#348)
Can bit timings are a bit in depth but aren't actually that complex,
especially as parametrized on the stm32 canbus modules. They depend on
the clock you're giving them, what you want the time quantum to be, what
you want the bitrate to be, and when you want to sample. Given all those
you can calculate numbers, and with c++ consteval it's easy to do it at
compile time with static asserts for out of range values (including
narrowing!). It's a bit messy because it all needs to be templates,
though.

The bit timings here are set to match the values that were passed to the hal initializer before the refactor (verified by unit tests).