fix(can): set 500kbaud exactly as bitrate
CanFD requires extremely close clock alignment for nominal bit times. It
can handle a certain degree of source-clock drift, but everything has to
agree ahead of time on what times it actually _wants_ to allow.

We can achieve a precisely 100ns time quantum duration on the
L5s (source clock: 110mhz off the pll), verdin (source clock: 20MHz),
and G4s (if we make the source clock 170MHz off the pll instead of 85
off the periphclk) and therefore get exact nominal agreement, which in
practice is well within the range CAN can handle.

In fact, it's so within that range that we can bump the bitrate to
500Kbaud easily.