feat(can): blink leds when there is a canbus error (#352)
* feat(common): add ISR interface to timers

This will be needed for canbus error state handling.

* feat(can): add bit_timings concept

Useful to recognize one of the templated structs.

* feat(can): LED can mirror protocol state

When the CANbus peripheral goes into an error state, we may be able to
try and send messages upstream, but they may not get through since the
peripheral is in an error state. What we can do, though, is make the LED
blink at different rates. This commit therefore

- Adds a timer to the canbus instance that is used to drive LEDs
- Adds interrupt callbacks through to the can bus cpp that handle FDCAN
  protocol level errors and end up changing the rate the LED flashes
- Add new ctors to specify LEDs, and a start() method that calls the c
  can_start() so that it can also start the timer at an appropriate time

* refactor(*): use LED to indicate can status

Use the LEDs to indicate can status instead of motor status, which
requires switching the LED instances over to the canbus constructor and
also using the cpp start method instead of the c one.