refactor(head): targets for head revisions
We now have two revisions of the head: proto, and rev1 (the first
production rev). They will have different settings and different source
files but otherwise be pretty similar. There's a lot of stuff we can do
exactly the same for both, so we can do that in a list foreach; stuff
that's different can remain open-coded.