fix(i2c): restrict the message queue size
In reality, message queues are restricted to a certain size. Some of the tests around polling relied
on the fact that dequeue (the simulation queue backing) does not have a set size. This PR restricts
that size to the default queue size (10) which also happens to be the poll count size.