refactor(tmc2160): modify how we scale current on the tmc2160 motor driver
The datasheet for the 2160 driver suggests that current be scaled based on the global scalar as
opposed to calculated the value of IRUN Or IHOLD directly. I think we probably need to dig a bit
deeper into stealthchop, whenever we want to actually configure it properly, because there were some
warnings around changing the global scalar while using this feature.