feat(gantry): support rev1 boards (#395)
* refactor(gantry): add revisions to cmake firmware

Add a list of revisions to the firmware cmakelist and refactor to build
targets for both revisions and both axes. By iterating through the axes
rather than writing down both targets, we can make things a little more compact.

* refactor(gantry): add revisions to core cmake

and use those revisions in firmware and simulators

* refactor(gantry): split tasks, interfaces, main

These things will all need to be different between revisions. For now,
get ahead of things by splitting implementations of tasks and interfaces
by revision, and removing the common queue types to a new queues header.

Main and interfaces need separate headers and source files, as do tasks.

* feat(gantry): bring up rev1 controllers

These are mostly the same but have a tmc2160 instead of a tmc2130, so we
swap that out in the rev1 code.