refactor(motor-control): Enable automatically
When the motors are about to move, they should get enabled without
requiring a previous enable request. This moves the enable motors
message into something used only for turning on motors and holding, as
in the OT-2, rather than a necessary step before turning motors -
removes the possibility for programmer error.