refactor(motor-control): setup automatically (#400)
We have this setup request canbus message that commands all the micros
to do a bunch of motor driver configuration that they have built in
anyway. Rather than require this, and therefore run the risk of a
programmer forgetting to do it, let's just make the setuprequest happen
internally without prompting.