feat(motor-control): encoders in hardware
Add encoders to the hardware interfaces.

The encoders now present data as int32_t numbers of pulses. This number
is not just the currently-held value in the timer registers, but also
the sum of the net number of overflows, which is handled by adding an
encoder_overflow callback that must be called from the timer overflow
interrupt.

The encoder timer system works by using quadrature encoders to drive a
timer up or downcounter. That means that at any given time, barring
overflows, the value currently held in the timer counter register
reflects the current absolute position of the encoder, as a uint16_t.

When that position overflows high or low, an interrupt occurs and the
counter register wraps over in the specified direction. We can turn that
into effectively the high 16 bits of a 32 bit counter - an overflow
positive is like the number going from 2^16-1 to 2^16, so we add 1 to
the high bits; an overflow negative is like the number going from 0 to
-1, so we subtract a bit and let the signed math do the work.

Once we've formed the signed upper part of the 32 bit number, we can add
the unsigned lower part to it, and everything magically works.

This all has to happen in the hardware interface because it needs to be
accessible from multiple threads. This is, in fact, bad and not the
right thing to do; but since these are machine words it should more or
less work, and the fix is pretty involved - this number (and the step
position) should be exposed via mailboxes or notifications.